#399 Extract geometric intersections calculation to utility class
There are currently two places in the code where a Direction object is
used to compute the intersection between a line that starts at the
center of a shape with the contour of the shape. These computations are
complex and related to geometry, so they would be best encapsulated
within a class of the geometry package, where their parallel structure
will be more evident. This will also facilitate improving diagram
rendering by having a well-defined place to put intersection
calculations for all shapes, so that we can add them for more complex
shapes (such as rounded rectangles and ellipses) later.