#416 Refactor edge constraints checking
`ConstraintSet` is now stored as a private static final field in each
diagram builder. This reduces object creation by eliminating the need to
create and merge new `ConstraintSet` objects each time `canAdd` is
called. For this reason, I removed the `merge` method in `ConstraintSet`
which is no longer necessary.
The `ConstraintSet` field can be accessed by method
`getEdgeConstraints()` which simply returns the set of constraints. I
see no encapsulation issue in doing so since there is no method
available to modify contents of a `ConstraintSet` which will be
returned.
 `Satisfied` takes in the `canAdd` parameters in order to check the
conditions whilst keeping the code simple.