#427 Remove node's position as properties
The x and y values for the position of a node are no longer managed as
node properties. This design decisions created a special case that adds
unnecessary complexity and prevents the Point class from being
immutable.

From this point on, the position of a node is solely managed through
using the position() and moveTo(...) methods on Node.