#439 Add TestPersistenceService2
This class adds a test that focuses specifically on testing that the
persistence service works, without verifying the details of the diagram,
which will be moved to specialized layout test classes.

The class is named ...2 to avoid a class in the revision history. The
intent is to rename it TestPersistenceService once the current class
with that name is reorganized.