#375 Add a NodeStorage, unit tests, and a performance test.
A NodeStorage was added to each NodeViewer to improve the performance of DiagramViewer.draw(diagram). For each call to NodeViewer.getBounds(node), if the NodeStorage is deactivated, then the bounds will be computed as returned. If the NodeStorage is activated, then it will first check whether the bounds have already been computed, and if so, it will return them directly. Otherwise, it will compute the bounds to be returned and store them for future reference. As soon as the NodeStorage is deactivated, it is cleared.

A performance test was added to notice the benefits of this change.

Unit tests are included to test the new behaviour.