Make the segmented edge label positioning without overlap consistent.
In order to avoid the label overlapping with the segments of a segmented edge, we introduce the method isStepUp which takes in an edge and returns a boolean representing whether the edge represents a visual step up or a step down. This information is relevant in the method getAttachmentPoint to unify the behaviour of the label's position with respect to vertical middle segment edges. As of now, there should be less overlap between the middle label of a segmented edge and its segments.