#375 Added Layouter, ClassDiagramViewer, and StoredEdgeViewer
In order for each diagram to be associated with its own unique storage,
ClassDiagramViewer has been converted into a stateful diagram viewer
which is unique to each class diagram .This way, each ClassDiagramViewer
aggregates an EdgeStorage object and a StoredEdgeViewer which is used to
draw stored edges. When drawing edges, ClassDiagramViewer is responsible
for invoking Layouter to plan the edges in the diagram.
ClassDiagramViewer also calls on its StoredEdgeViewer constituent to
draw the diagram’s edges using EdgeStorage.

DiagramType maps diagrams to their DiagramViewer which is accessible
using the public method viewerFor(Diagram
pDiagram). DiagramType also maintains a map of class diagrams and their
ClassDiagramViewers. So, viewerFor(Diagram pDiagram) returns the unique
ClassDiagramViewer for pDiagram if it
is a class diagram.

Each ClassDiagramViewer aggregates a Layouter object which is
responsible for planning and storing the paths
of edges in the diagram. This features is invoked through Layouter’s
only public method: layout(pDiagram).

Add tests for ClassDiagramViewer, StoredEdgeviewer, and Layouter.