#375 Use EdgeStorage to draw selection handles and remove unused imports
Add a method drawSelectionHandles(pSelected, pGraphics) to
DiagramCanvas. This method is called by paintPanel() and checks if the
current diagram is a class diagram. If so, the method delegates the draw
selection handles functionality to ClassDiagramViewer.
ClassDiagramViewer uses EdgeStorage to get the start and end connection
points of a stored edge when drawing the selection handles.

Remove unused or duplicate imports.