Solve Issue #500
For simplicity, the solution does not set up a stage just to serve
as a parent for the Alert dialog showing error on file open from the
command line. The only consequence is that if the dialog appears, a
default "Java" icon will be shown in the OS taskbar instead of the
JetUML-branded icon.